core: Move intrinsic header to core
