+++ libs/core/include/core_intrinsic.h
