+++ libs/core/src/bitset.c
@@ -1,10 +1,9 @@
 #include "core_bits.h"
 #include "core_bitset.h"
 #include "core_diag.h"
+#include "core_intrinsic.h"
 #include "core_math.h"
 #include "core_types.h"
-
-#include "intrinsic_internal.h"
 
 usize bitset_size(const BitSet bits) { return bytes_to_bits(bits.size); }
