+++ libs/core/src/math.c
@@ -1,8 +1,7 @@
 #include "core_array.h"
 #include "core_diag.h"
+#include "core_intrinsic.h"
 #include "core_math.h"
-
-#include "intrinsic_internal.h"
 
 u64 math_pow10_u64(const u8 val) {
   static const u64 g_table[] = {
