+++ libs/geo/src/quat.c
@@ -1,11 +1,10 @@
 #include "core_diag.h"
 #include "core_float.h"
+#include "core_intrinsic.h"
 #include "core_math.h"
 #include "geo_matrix.h"
 #include "geo_quat.h"
 #include "geo_vector.h"
-
-#include "intrinsic_internal.h"
 
 #define geo_quat_simd_enable 1
