+++ libs/geo/src/vector.c
@@ -1,11 +1,10 @@
 #include "core_bits.h"
 #include "core_diag.h"
 #include "core_float.h"
+#include "core_intrinsic.h"
 #include "core_math.h"
 #include "core_rng.h"
 #include "geo_vector.h"
-
-#include "intrinsic_internal.h"
 
 #define geo_vec_simd_enable 1
